Restructure test modules
